factory for use case
